Moved import operations from main to mgui. Separated import MGUI from old magicgui.
